Chapter4-9 章末問題アップデート

+++ Chapter4-9 章末問題/List7_Ex01.c
@@ -12,7 +12,7 @@
 
 int main(void){
     int fd;                             //ファイルディスクリプタ
-    float temp;                         //温度 実数値℃
+    double temp;                         //温度 実数値℃
     fd = wiringPiI2CSetup(ADT7410_ADR); //温度センサのI2Cセットアップ
     while(1){
         temp=Adt7410Read(fd);
